основной сервис

+++ stats-module/stats_dto/src/main/java/ru/practicum/dto/StatsDto.java
@@ -12,5 +12,5 @@ public class StatsDto {
 
     private String app;
     private String uri;
-    private Long hits;
+    private Integer hits;
 }
